Maintain equivalence classes of streams (based on escape information). If any stream in the class is closed, then they are all considered closed.  This should get rid of some false positives.
We should definitely revisit this in the future: for example,
if you wrap a stream with a buffered output stream, and forget
to flush it, then if you close the wrapped stream, you lose data.
A similar problem can occur for input streams.  We have all of the
information we need to handle this, so we just need to do
a bit more work with the escape information.


git-svn-id: https://findbugs.googlecode.com/svn/trunk@2636 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3