Bill has been a busy boy.
Fixed typos in messages.

Made DLS a style warning.

Fixed some problems with how DLS was being computed.

Added a check for an integer value being cast to a double and then passed
to Math.ceil.

Added toString methods for WarningPropertySets

Rewrote and simplified code to check for result of
integer division being cast to double

Expanded check for SQL injection. This now could catch to much, and we need to
add some data flow analysis to make it smarter.


git-svn-id: https://findbugs.googlecode.com/svn/trunk@3782 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3