[ 1632009 ] Eclipse plugin context menu fix for eclipsePlugin1.1.3
Hi, 

I am sure others have noticed that the Eclipse plugin's context sensitive menu is, well, not context sensitive.

It always clears bug markers for the entire project and only works either with projects that use the same folders for source and class files or with other projects at the project level only.

The problem is due to an assumption in the plugin, that class files will always be present in the same location as the .java files. Which is not true for most projects with the least amount of complexity.

So I wrote this fix for the plugin that does both, fixes the 'Clear Bug Markers' menu item so that markers are only removed for the context (i.e. for the file, package, source folder or project where the the context menu was launched in the package explorer) and the 'Find Bugs' menu item finds (i.e analyses) only the files under the context.

In other words, you can now find bugs and clear markers for an individual file, package, source folder or the entire project.

This is especially important for projects with a large number of packages and source files, where running findbugs on the entire project just takes too long.

I have taken extra effort to make sure that all kinds of Java projects are covered (even projects with custom output locations for different source folders/packages and externally linked projects), and done in the most efficient and unobtrusive way possible that I could think.

Please feel free to ask any questions and provide feedback on this patch.

Hopefully this patch will make it soon to the HEAD for version 1.1.4.

Thanks,
Dhiraj.



git-svn-id: https://findbugs.googlecode.com/svn/trunk@7533 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3