Multithreaded correctness - does not release lock on all exc
The following code shows the bug;

Method does not release lock on all exception paths

By replacing the lock code with the commented out code, no bugs are
reported.

I'm not sure how just using the lockArray instead of a reference to an
element of the lockArray could create a situation where an exception could
be thrown that couldn't be in the first case.

git-svn-id: https://findbugs.googlecode.com/svn/trunk@9952 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3