Apply patch 2829771 Various build fixes
A bunch of things rolled together.

1. Remove the bin directory from source control and add to svn:ignore. Move
bin/findbugs.bat to src/scripts/standard/findbugs.bat (can run dos2unix on
it so it is nicely editable for everyone). Delete bin on clean, and when
building, fix up CRLFs on *.bat.

2. Set an explicit encoding for src/junit since two files use á in
ISO-8859-1 encoding, and you get a warning from the compiler if your system
encoding is UTF-8 (true of a typical Linux installation).

3. Use includeantruntime="false" to avoid accidentally dumping Ant JARs in
your compiler classpath. Introduce ${ant.core.lib} (i.e. ant.jar) in the
one place it is actually wanted.

4. Fix two source files which do not compile under JDK 7's javac. See
http://www.netbeans.org/nonav/issues/show_bug.cgi?id=163763 for background.
The fix is innocuous in this case, but if after reading the description
given above you still think the code ought to compile according to the JLS,
best to file it on bugs.sun.com.

BTW it is also possible to just delete the nbproject directory if you use
http://wiki.netbeans.org/AutomaticProjects to work with the project in
NetBeans. Just install that plugin, use File > New Project > Java >
Automatic Project to start, select FB source tree root, and when opened
press F11 to build and pick a target (default of 'build' seems to work
well). No need to maintain any NB-specific "freeform" project metadata (the
current version of which is in fact wrong).



git-svn-id: https://findbugs.googlecode.com/svn/trunk@11526 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3