Apply patch 2807092 Extended start script for Windows
Please find attached a start script to perform various FindBugs features
from Command line. There are a number of scripts available to run on Unix,
but those are not available for Windows. I have modified existing
findbugs.bat to be able to run same functions are available with Unix
scripts.

The difference is: I did not create all scripts separately, but added
command line options handling into one script. Different start class is
selected depending on the options. Here are the options:

-addMessages (edu.umd.cs.findbugs.AddMessages)
-computeBugHistory (edu.umd.cs.findbugs.workflow.Update)
-xmltotext (edu.umd.cs.findbugs.PrintingBugReporter)
-copyBS (edu.umd.cs.findbugs.workflow.CopyBuggySource)
-defectDensity (edu.umd.cs.findbugs.workflow.DefectDensity)
-filterBugs (edu.umd.cs.findbugs.workflow.Filter)
-listBugDatabaseInfo (edu.umd.cs.findbugs.workflow.ListBugDatabaseInfo)
-mineBugHistory (edu.umd.cs.findbugs.workflow.MineBugHistory)
-printAppVersion (edu.umd.cs.findbugs.workflow.PrintAppVersion)
-printClass (edu.umd.cs.findbugs.workflow.PrintClass)
-rejar (edu.umd.cs.findbugs.workflow.RejarClassesForAnalysis)
-setInfo (edu.umd.cs.findbugs.workflow.SetBugDatabaseInfo)
-unionBugs (edu.umd.cs.findbugs.workflow.UnionResults)
-xpathFind (edu.umd.cs.findbugs.workflow.XPathFind)

git-svn-id: https://findbugs.googlecode.com/svn/trunk@11527 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3