Huge refactoring to enable running the appengine cloud as a standalone application using Jetty, HSQLDB, and DataNucleus. It seems to work!
The main differences between the codebases (appEngineCloud vs localCloud) are:
- App Engine uses a proprietary Key object

Unfortunately there's a bit of code duplication necessary to get this working. This is because while JDO does allow for inheritance, the base class needs to provide the primary key - for some weird reason - and since in a few cases a Key is the primary key, we can't use a common base class. The web.xml is also duplicated, not sure if there's a way to share common aspects of a web.xml.

Still to do:
- Make sure it builds in Eclipse
- Create two sets of unit tests
- Make Ant file to build it from command line
- Make sure the HSQLDB saves and loads from disk correctly

git-svn-id: https://findbugs.googlecode.com/svn/trunk@11936 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3