Fix for the deadlock/bug from http://sourceforge.net/tracker/?func=detail&atid=614693&aid=3312970&group_id=96405
Looks like the one lock in question is at line 182 in TypeQualifierValue (instanceof of InheritableThreadLocal):

    private static final InheritableThreadLocal<AtomicBoolean> performingValidation
        = new InheritableThreadLocal<AtomicBoolean>() {
        @Override protected synchronized AtomicBoolean initialValue() {
            return new AtomicBoolean();
        }

    };

and the problematic code is the initialValue() method, which overrides the super method BUT adds "synchronized" to it. Was introduced at 2010-10-14  (see http://code.google.com/p/findbugs-hg/source/detail?r=d9ed0ae42d646f6176b0f9d174d0708c2eb90402).

The deadlock seems to happen if there are two different FB threads (== Eclipse jobs) running in the same VM.

One job is already analyzing the code and takes the lock on  TypeQualifierValue$2 by calling TypeQualifierValue.performingValidation.get();. After that, inside the ThreadLocal.get() method something happens (some class need to be loaded?) which needs a lock on the current classloader (DefaultClassLoader). So it will wait until DefaultClassLoader lock will be freed.

Another job is already loading class (which goes to the synchronized method ClassLoader.loadClass() which calls parent classloader (DefaultClassLoader) to load class).  So here is the first lock taken on the DefaultClassLoader object. Then somewhere during the classloader code the ValidationSecurityManager is asked to validate the class and waits for the first thread which is holding the lock in TypeQualifierValue$2 code (this locks down the ValidationSecurityManager due the lock on synchronized initialValue() call).

So the fix is to simply to remove the "synchronized" keyword from the initialValue() call.

git-svn-id: https://findbugs.googlecode.com/svn/trunk@13265 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3