#282 Add option to read arguments from file.
This patch adds two command-line flags: "-auxclasspathFromFile" and
"-analyzeFromFile". These flags instruct FindBugs to read the
auxclasspath or the analysis targets from a file, rather than from the
command-line or standard input.

The point of these changes is that currently a user cannot specify both
"-auxclasspathFromInput" and "-xargs" in the same command. They cannot
both use stdin. These changes are useful when both the auxclasspath and
the analysis targets are too large to fit on the command line.

Contributed by "JVS" user.