[PATCH] Added support for run-time-customized labels (v2)
From: Kevin Lubick <kjlubick@ncsu.edu>, see
https://sourceforge.net/p/findbugs/feature-requests/309/

If you extend CustomLabelBugResolution, the base label can have a
placeholder like `replace with YYY.isNaN()` and the `YYY` part will be
substituted out with the what the passed-in visitor gives.

In the above example, a simpile visitor might detect if `Float` or
`Double` will be used to make the replacement and give the user a more
accurate idea of what change will be made.

Documentation and licenses have been added, including some null-related
annotations.