Overriding multifix
This somewhat fixes the problem of multiple fixes in one file.
However, and I think this is because multiple rewrites are getting dished
out, only one bug gets resolved at a time.  But, a careful stepthrough
with the undo mechanism indicates they both are done, just one undoes the
other.