Fully resolves Request #277 - custom colors
Resources that were useful to finding the extension point to modify:

http://help.eclipse.org/kepler/index.jsp?topic=%2Forg.eclipse.platform.doc.isv%2Freference%2Fextension-points%2Forg_eclipse_ui_editors_markerAnnotationSpecification.html
http://cubussapiens.hu/2011/05/custom-markers-and-annotations-the-bright-side-of-eclipse/

The OverviewRuler is the name of the side bar that has the colored patches
for warnings, errors, bookmarks, etc.

Basically, there are 4 annotations defined extending
org.eclipse.ui.editors.annotationTypes, which by default, are formatted
like the standard warnings.  By overriding the 4 fields
(colorPreferenceKey,colorPreferenceValue,overviewRulerPreferenceKey and
textPreferenceKey) in the markerAnnotationSpecification extension, I was
able to establish both a custom color, as well as make the user be able to
specify a color, if they choose.

For good measure, I specified the rest of the fields (like
overviewRulerPreferenceValue) to make sure they showed up like warnings
(as some of these defaulted to false).