Trying a more memory efficient caching
The old way didn't like keeping many many compilation units in memory.

This modification allowed for 1300+ simultaneous quickfixes (I did have
to boost the memory given to Eclipse to 1G to handle the fixes, which is
reasonable for 1K+ fixes)

Still want to disable automatic scanning until after the fixes are done.