`FindBugsViewerTask`: use default look and feel by default
Change the default value of the `look` property in `FindBugsViewerTask` from
`"native"` to `null`, meaning the default L&F is used.

`GUI2CommandLine` only allows the options `plastic`, `gtk` and `native`; it is
possible to get the default L&F by specifying `look=""`, for example, but that
causes `GUI2CommandLine` to emit an error.

This change also makes the default L&F match what you get when running
SpotBugs standalone.