update gradle plugin to consistently use SpotBugs
Many places in the gradle plugin were still referring to
FindBugs. This can cause some confusion, in particular,
since the default gradle FindBugs plugin will also likely
be in the classpath. Some of the links in javadoc comments
have also been updated.

The `AbstractCodeQualityPlugin` class has been removed
because it wasn't being used. The code refers to the one
provided by gradle rather than the local copy.

Note, this changes the name of the tasks to have a prefix
of `spotbugs` instead of `findbugs` (see test case).