Don't remember module-info seen to avoid popping up silly BCEL errors
module-info is evil and should ideally not exist as a class file at all,
but it does. So if we see a reference to it because we are trying to
read the type from the module-info.class file, we should not remember
its class descriptor to avoid being later used as one of referenced
classes to analyze. It is not useful, it cannot be referenced, it is
just pain.

Added extra test case for issue #408 where the module-info resides in
the same package as another class to analyze. Additionally fixed the
test driver to allow this test run inside Eclipse, where all classes
reside in same class folder.