497: Do not print exceptions for unsupported classpath files
Assume that any files thought to be a zip file in the classpath that
cannot be loaded as a zip file should be ignored. Many files may be
placed on the classpath but this does not guarantee they are valid.

Prior to this change, an exception would be thrown which was presented
to the user regarding the invalid path. This did not cause a failure,
but was hard to discern from an actual exception. Ignoring these files
reduces unnecessary output and mimics the behavior of the JVM which does not
complain about invalid files on the classpath.