fix: fix SARIF1011 reported by SARIF validator
$schema: The  property value https://schemastore.azurewebsites.net/schemas/json/sarif-2.1.0-rtm.4.json
does not refer to the final version of the SARIF 2.1.0 schema. If you are using an earlier version
of the SARIF format, consider upgrading your analysis tool to produce the final version.
If this file does in fact conform to the final version of the schema, upgrade the tool
to populate the  property with a URL that refers to the final version of the schema.