Bumped jaxen from 1.1.6 to 1.2.0 version
This fixes issue #1316 and finally fixes issue #1315.

This change switches from Jaxen 1.1.6 to 1.2.0 to allow compilation of
the projects that rely on SpotBugs project with Java 11.

Jaxen is a transitive optional dependency of dom4j that we use in
SpotBugs.

Jaxen 1.1.6 ships org.w3c.dom package, which is also now shipped by JDK
in java.xml module. Java 9+ comiler disallow compilation if modules on
module path ship same package. So compilation of spotbugsTestCases with
Java 11 fails if the target release is 11 with the error: "The package
org.w3c.dom is accessible from more than one module: <unnamed>,
java.xml".

Solution is to update Jaxen to 1.2.0 version, that does not ship
org.w3c.dom package anymore.

Signed-off-by: Andrey Loskutov <loskutov@gmx.de>