Improve FindReturnRef to also warn in case of shallow copied arrays
Some programmers try to avoid returning mutable private members by
returning thir clones. However, in case of array the clone() method does
a shallow-copy instead of a deep one. If the array members are of a
mutable type, then we still return mutable private data. Similarly, when
passing an array to a method, the method should clone its members
one-by-one, if they are of a mutable type, instead of cloning (thus
shallowly copying) the array itself. This patch improves the
`FindReturnRef` detector to also issue `EI`, `EI2` and `MS` warnings in
these cases.