Add new rule PERM_SUPER_NOT_CALLED_IN_GETPERMISSIONS
SEI CERT rule SEC07-J requires that the superclass's getPermissions() is
called when writing a custom class loader. This method is defined in
class java.net.URLClassLoader. Thus any class inheriting directly or
indirectly form URLClassLoader and overriding its getPermissions()
method must call the superclass's getPermissions() method in this
method. It is not enoguh to just call it but the the return value (an
instance of java.security.PermissionCollection) must be initialized by
calling the superclass's method instead of creating a new instance from
scratch.

This patch adds a new detector to check this rule.