Detector for bad End of Stream Checks
java.io.FileInputStream.read() and java.io.FileReader.read() both return an integer.
The value of this integer is -1 if the end of the stream is reached. Since
FileInputStream reads bytes and FileReader reads characters their return value
must be casted to these types. However, if the casting happens before the check
for -1 then -1 becomes indistinguishible from 0xFF in case of FileInputStream and
becomes Character.MAX_VALUE (the highest positive value of a character) in case
of FileReader. SEI CERT rule FIO08-J describes these bugs in detail. This patch adds
a new error message EOS_BAD_END_OF_STREAM_CHECK and a new detector
(FindBadEndOfStreamCheck) which checks for this kind of error.