Add option to user-defined list of immutable classes
Although our heuristics considers a class as mutable only in the case it has a setter, some users may want to handle some of them as immutable in their code where security is less important than perofrmance. This patch allows them to define a comma-separated list of classes which are considered as immutable. Whole packages may also be whitelisted using an asterix for the class name.