Exclude methods returning any object from potential setters when deciding on the mutability of a class
Methods which returned an instance of the owning class itself are probably no setters because they usually return a new instance instead of modifying the existing one. The algorithm did care for that, but it turned out that in many cases this is not enough. Some immutable classes contain methods with a setter-like name which return an instance of a class which is different from their owning class. This patch changes the behavior of the alrogithm to skip these methods as well.