New detector for checking SEI CERT rule SEC02-J: Do not base security checks on untrusted sources
A new detector called FindPotentialSecurityCheckBasedOnUntrustedSource approximately checks for
SEI CERT rule SEC02-J: Do not base security checks on untrusted sources. Specifically, it checks for the
scenario presented in the example code: the programmer calls a non-final method of a non-final class
instance received as a parameter of a public method of a public class to do a security check before
entering a doPrivileged block and then calls the same method inside the block. Since the method is
public it can be called by anyone and because the class of the parameter type is non-final, neither
the method is final which is called before and inside the doPrivileged it could have been overridden
by the caller to bypass the security check by returning different values on the first and subsequent
invocations.