Remove \u200b from FindOverridableMethodCall.visitBootstrapMethods()
The method ends with \u200b unicode character and Eclipse compiler
assumes that this method can't override the visitBootstrapMethods
defined in the
org.apache.bcel.classfile.Visitor.visitBootstrapMethods(BootstrapMethods)

Since this character is invisible, it was added by mistake and should be
removed to allow SpotBugs compilation in Eclipse.

This fixes issue #1903.