Change ValueRangeAnalysis to a data-flow analysis
Currently, SpotBugs includes a value-range analysis which is only used
for detecting redundant conditions. However, it could be used for
detecting other kind of errors, such as arithmetic errors as well. This
PR refactors this analysis to a data-flow based one while keeping the
functionality of the existing redundant condition analysis based on top
of it.