Add new rule PERM_SUPER_NOT_CALLED_IN_GETPERMISSIONS (#1537)
* Add new rule PERM_SUPER_NOT_CALLED_IN_GETPERMISSIONS

SEI CERT rule SEC07-J requires that the superclass's getPermissions() is
called when writing a custom class loader. This method is defined in
class java.net.URLClassLoader. Thus any class inheriting directly or
indirectly form URLClassLoader and overriding its getPermissions()
method must call the superclass's getPermissions() method in this
method. It is not enoguh to just call it but the the return value (an
instance of java.security.PermissionCollection) must be initialized by
calling the superclass's method instead of creating a new instance from
scratch.

This patch adds a new detector to check this rule.

* Debug printouts removed and other fixes

* Header comment fixed from FindBugs to SpotBugs

* Updated based on the review comments by @KengoTODA

* Updated according to the comments of @KengoTODA

* Grammar fix

* Checking for all ancestors instead of the direct superclass

* Minor fix

* Changelog updated after the merge

* Update spotbugs/src/main/java/edu/umd/cs/findbugs/detect/PermissionsSuper.java

Co-authored-by: Kengo TODA <skypencil+github@gmail.com>

* Merge `master` into `PermissionsSuper`

Co-authored-by: Kengo TODA <skypencil+github@gmail.com>