Fix for Issue 2040
This PR fixes issue #2040 by limiting the errors reported by the detector `ThrowingExceptions` to cases where
`java.lang.Exception` or `lava.lang.Throwable` appears in the exception specification of th method but it is neither
inherited from an overridden method nor is is coming from another method that the method invokes. Syntathic
methods are also omitted as well as methods which throw generic exceptions.