New bug type: AA_ASSERTION_OF_ARGUMENTS
Using assertions to validate arguments of public functions is a bad and dangerous practice because the validation does not happen if assertions are disabled, thus the behavior of the program depends on this compilation option. Therefore, we introduced a new detector `FindArgumentAssertions`  to detect such cases and report `AA_ASSERTION_OFR_ARGUMENTS` for each of them.

Co-authored-by: Gábor Kutas <@vodorok>