New bug type: ASE_ASSERTION_WITH_SIDE_EFFECT
Assertions must never have side effects because they make the behavior of the program dependent on the option whether assertions are enabled or disabled. Therefore we introduced a new detector `FindAssertionsWithSideEffects` to detect such cases and report `ASE_ASSERTION_WITH_SIDE_EFFECT` for each of them. Assertions which contain a method call are also checked andd `ASE_ASSERTION_WITH_SIDE_EFFECT_METHOD` is reported on them if the called method looks like a setter and the object is an instance of a mutable class.

Co-authored-by: Gábor Kutas <@vodorok>