Fix detector FindSqlInjection to detect bug SQL_NONCONSTANT_STRING_PASSED_TO_EXECUTE with high priority also in Java 11 and above
Instead of using `StringBuffer` or `StringBuilder` internally, Java 11 and above uses a dynamic call to `makeConcatWithConstants()` to append strings. However, in case of loops, this is still less efficient than creating a `StringBuffer` or `StringBuilder` outside the loop and using it inside to build the string. Previously, the `FindSqlInjection` detector assigned low priority in Java 11 and higher in the same cases where it assigned high priority in Java 8. This PR fixes this. See issue [#2183](https://github.com/spotbugs/spotbugs/issues/2183).