Improve compatibility with later version of jdk (>= 13)
A method with signature **newFileSystem(Path path, Map<String,?> env)**
was introduced in version 13, which results in the conflict in
SourceFinder.java:361. According to FileSystems.java in version 8
(which is recommended to compile this project), **newFileSystem(Path path, ClassLoader loader)** should be used in this context. So an explicit cast should be used here.