Fix OpcodeStack to handle propagation of taints properly in case of string concatenation in Java 9 and above
Instead of using StringBuffer or StringBuilder internally, Java 11 and above uses a dynamic call to makeConcatWithConstants() to append strings. Previously, `OpcodeStackDetector` did not handle the taint propagation properly in case of this dyanamic call which led to false negative such as the one described in issue [#2184](https://github.com/spotbugs/spotbugs/issues/2184). This PR fixes such issues by adding code to `OpcodeStackDetector` to handle this case as well.