Fix detector StringConcatenation to detect bug SBSC_USE_STRINGBUFFER_CONCATENATION also in Java 11 and above (#2186)
* Test for Issue #2182

* Fix detector `StringConcatenation` to detect bug `SBSC_USE_STRINGBUFFER_CONCATENATION` also in Java 11 and above

Instead of using `StringBuffer` or `StringBuilder` internally, Java 11 and above uses a dynamic call to `makeConcatWithConstants()` to append strings. However, in case of loops, this is still less efficient than creating a `StringBuffer` or `StringBuilder` outside the loop and using it inside to build the string. Previously, the `StringConcatenation` detector missed all the cases in Java 11 and higher, but this PR fixes this. See issue [#2182](https://github.com/spotbugs/spotbugs/issues/2182).

* Updated according to the comments of @ThrawnCA

Co-authored-by: Kengo TODA <skypencil@gmail.com>