Fix Issue #184 (#2295)
* Fix Issue #184

The logic conditionals were entangled in an invalid way that produced
nonsense debug log messages. Simplified the logic and reduced the logic
branches to make the operation direct and thus fix the problem.

* Added unit test to prove correct behavior of priority and rank

This unit test serves as proof that the reportBug function is operating
as expected with respect to reporting based on priority and rank
threshold values.

As a bonus this test also verifies that the relaxed mode will override
any rank or priority and always report bugs found.