New bug type: ASE_ASSERTION_WITH_SIDE_EFFECT (#2130)
* New bug type: ASE_ASSERTION_WITH_SIDE_EFFECT

Assertions must never have side effects because they make the behavior of the program dependent on the option whether assertions are enabled or disabled. Therefore we introduced a new detector `FindAssertionsWithSideEffects` to detect such cases and report `ASE_ASSERTION_WITH_SIDE_EFFECT` for each of them. Assertions which contain a method call are also checked andd `ASE_ASSERTION_WITH_SIDE_EFFECT_METHOD` is reported on them if the called method looks like a setter and the object is an instance of a mutable class.

Co-authored-by: Gábor Kutas <@vodorok>

* Typo fixed

* Updated according to the comments of @KengoTODA

* Typo fixed

* reorder arguments to match method signature, fix typo

* add license header

* CHANGELOG updated

* Fix MutableClasses.looksLikeASetter nullsafe

---------

Co-authored-by: Ádám Balogh <adam.balogh@ericsson.com>
Co-authored-by: Judit Knoll <judit.knoll@sigmatechnology.com>