[junit] Apply junit 5 to code that handles junit 3/4
Discovered 'DontAssertInstanceofInTests' when upgrading to junit 5.  The others seem to imply same need.  While 'TestaseDetector' probably is a no-op, its better to mention both.  The dumbMethods seems to need the same check or would benefit.  And finally just noting on FindNullDeref that its not needed for junit 5 given junit 5 doesn't do it that way.