[tests] Add handling around our packaged java tests
for java11 or better, add flag to ignore java 8
for java17 or better, add flag to ignore java 8/11
for test that seen failing with java 11, add ignore on java 8/11

Note: We are going to explore showing still building with java 8 separately so this preps based on our coding style here.  It does not mean java 8 will work.  Further, the reason we are doing this effort is for contributors so they are not scared away by not having a java version to use to test as many are stuck at various points and we get more engagement from developers if they are not so limited.  That is a separate effort bot noting as the PR itself has comments in this regard.