Migrate to full junit 5 and add some associated code expected to support junit 5 - fixes #2596 (#2605)
* [junit5] Convert to junit 5

* [junit] Apply junit 5 to code that handles junit 3/4

Discovered 'DontAssertInstanceofInTests' when upgrading to junit 5.  The others seem to imply same need.  While 'TestaseDetector' probably is a no-op, its better to mention both.  The dumbMethods seems to need the same check or would benefit.  And finally just noting on FindNullDeref that its not needed for junit 5 given junit 5 doesn't do it that way.

* [ci] Update change log for junit 5 upgrade

* [ci] Update PR on changelog for junit5

* [tests] Per review, cleanup rest of left over publics from tests in other packages

* [tests] Add handling around our packaged java tests

for java11 or better, add flag to ignore java 8
for java17 or better, add flag to ignore java 8/11
for test that seen failing with java 11, add ignore on java 8/11

Note: We are going to explore showing still building with java 8 separately so this preps based on our coding style here.  It does not mean java 8 will work.  Further, the reason we are doing this effort is for contributors so they are not scared away by not having a java version to use to test as many are stuck at various points and we get more engagement from developers if they are not so limited.  That is a separate effort bot noting as the PR itself has comments in this regard.

* [ci] formatting