Support negated onlyAnalyze items
Having the ability to exclude only specific files is useful where a
project needs to be partially analyzed and it is the exclusions that are
known, not the inclusions.

Update -onlyAnalyze to recognize items starting with ! as exclusions,
allowing a lot more flexibility in specification of which classes are
included.

Signed-off-by: Robert Varga <robert.varga@pantheon.tech>