BetterCFGBuilder2 error when analyzing Spring AOT bytecode (#3064)
* test: sample reproducing the crash reported in issue #3059

* fix: Let `BetterCFGBuilder2.isPEI` handle `dup2` bytecode

The 'depth' variable keeps track of how many stack elements we need to
track back to find the reference of putfield. We start with 2 and
increase/decrease it when we see preceding instructions popping/pushing
elements.
When reaching a dup2 and depth comes to zero it means that the
instructions preceding the putfield have cancelled each other (the stack
is the same as before these instructions), so we reset depth to 2.