fix: use SignatureParser to extract the arguments of a method
ClassInfo.Builder.arguments() was incorrectly extracting the arguments
from a method signature: for instance the output for "(Xyz)V" was "("
instead of the expected "Xyz".
As a result the ClassInfo.Builder.build() was incorrectly matching
methods with same name (but different arguments) as bridge methods. This
caused errors ("Can't push void") in
TypeFrameModelingVisitor.pushValue() when the mismatched method had a
void return type.