ci: disable the SpotBugs analysis on sample code to speed up the build
The analysis of the sample code takes a long time and runs with an older
version of SpotBugs, so it occasionally crashes when a new sample was
added and the sample was reproducing a crash.
The analysis produces a 'spotbugsMain' file containing all the
dependencies and this is used in the integration test, replace that by a
custom task.