Disable Eclipse plugin signing to fix plugin build & release process
Originally plugin signing introduced in
f45d1917d8683b7ed10dbb581edd36a6a7c04a7e, now the key is expired and it
doesn't make sense to sign with expired key.

Not signed plugins are not nice (warning is shown on installation in
Eclipse) but same warning is shown if the key is not valid anymore and
the installation is still possible.

See https://github.com/spotbugs/spotbugs/issues/3406