[gha] Add a timeout to the build and cancel builds when new ones come in
Github now shows 'cost' that our runs take which at least for my personal account is the highest project I work on.  One of the issues on spotbugs is when we actively work fast pushing things and many builds are running at same time, this will cause the prior builds to stop when same branch and ref changes, so we still get multi platform testing without issue.  The only downside here is that we will get more cancels that appear as failed builds.  I've had this working on the maven plugin for few weeks and it has improved needless runs.

The timeout isn't a default used by github, I had a rouge unit test on one project that got stuck with java 24 and 25, I had been ignoring that for months as I didn't have time to look into it.  After 'costs' were shown, that bubbled up quick for a rarely used product showing costs could go into thousands quickly when in that case renovate was causing the job executions.  So timeouts added here, I put the buil at 60 minutes because it seems to vary, the others are fast so those are at 30 minutes.