Make sure renovate pins github action digests to protect against supply chain hacks
github is suggesting to do this now as supply chain attacks keep growing.  The prior usage would use something like v5 which is a mutable tag.  That by itself lends to supply chain attacks very well and is no longer suggested.  While we could point to a specific version, those are again tags and same issue.  Therefore its highly suggested to pin and renovate will show what top level version that is attached to.  It is then important for such renovate updates that we look to see what changed to ensure we don't merge a supply chain attack.