Always read latest preferences state from the disk on build in Eclipse
For the rest of the code it isn't that important (and constant reading
can affect performance), but every analysis execution should make sure
it is working with up-to-date settings.

Additionally removed unneeded call to getUserPreferences() from
readBugCollectionAndProject(), it was not needed anymore after one of
the previous code changes.

Fixes https://github.com/spotbugs/spotbugs/issues/3740