chore(build): Move all test modules to java 17 source compatiblity
We only support java 17 or better for building now and junit moved to java 17.  In order to pick that up, we should allow our test modules that are only used to build to adhere to the version we are required to use at this point.  This does not impact the runtime components that are all set to 11.