core(cleanup): Switch javax to jakarta on 'Nonnull' to further decouple jsr305 (#3858)
* core(docs): Move priority deprecation down to the other section with lots of deprecation removals

* core(cleanup): Migrate most 'javax.annotation.Nonnull' to 'jakarta.annotation.Nonnull'

efforts to get off jsr305.  with jspecify not fully supported here yet, this is most natural move now that we support commons annotations.  The move here is not everything.  Any place where 'When' was used, those remain, any where we use jsr305 inside code remains.  The changes were are our usage on this repo only which is substantial.  To keep the diff easy to follow, only the one import is changed.  Subsequent pulls will address others in same way.

* cores(docs): Update change log